Shrink header indexes to 32 bits to maintain consistent max length

+++ include/aws/http/private/hpack.h
@@ -34,15 +34,15 @@ void aws_hpack_static_table_clean_up(void);
 
 /* General HPACK API */
 AWS_HTTP_API
-struct aws_hpack_context *aws_hpack_context_new(struct aws_allocator *allocator, size_t max_dynamic_elements);
+struct aws_hpack_context *aws_hpack_context_new(struct aws_allocator *allocator, uint32_t max_dynamic_elements);
 AWS_HTTP_API
 void aws_hpack_context_destroy(struct aws_hpack_context *context);
 
 AWS_HTTP_API
-const struct aws_http_header *aws_hpack_get_header(const struct aws_hpack_context *context, uint64_t index);
+const struct aws_http_header *aws_hpack_get_header(const struct aws_hpack_context *context, uint32_t index);
 /* A return value of 0 indicates that the header wasn't found */
 AWS_HTTP_API
-uint64_t aws_hpack_find_index(
+uint32_t aws_hpack_find_index(
     const struct aws_hpack_context *context,
     const struct aws_http_header *header,
     bool *found_value);
@@ -51,7 +51,7 @@ AWS_HTTP_API
 int aws_hpack_insert_header(struct aws_hpack_context *context, const struct aws_http_header *header);
 
 AWS_HTTP_API
-int aws_hpack_resize_dynamic_table(struct aws_hpack_context *context, size_t new_max_elements);
+int aws_hpack_resize_dynamic_table(struct aws_hpack_context *context, uint32_t new_max_elements);
 
 /* Public for testing purposes */
 AWS_HTTP_API
